+++ Blurred-ios/ChannelViewController.swift
@@ -196,7 +196,7 @@ class ChannelViewController: UIViewController, UINavigationControllerDelegate, U
                 if let parseJSON = json {
                     let username: String? = parseJSON["username"] as? String
                     let name: String? = parseJSON["name"] as? String
-                    let imageUrl: String? = parseJSON["image_url"] as? String
+                    let imageUrl: String? = parseJSON["avatar_url"] as? String
                     let followerCount: Int? = parseJSON["followers_count"] as? Int
                     let followingCount: Int? = parseJSON["following_count"] as? Int
                     let bio: String? = parseJSON["bio"] as? String