+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 * Added collection view to channel view controller (doesn't do anything yet)
 * Added collection view to other channel view controller (doesn't do anything yet)
 * Moved profile divider line down by 5px
-* Added thumbnails to videos on people's channels (not other channels just your own for now)
+* Added thumbnails to videos on your own channel
+* Updated api request for avatar urls.
+* Added thumbnails to videos on others' channels (Thumbnail sizing broken)
 ## v0.2.5 (PRERELEASE)
 * Account settings editing now works (actually contacts the api)
 * Fixed bug where timer would stop forever and load any user info until app restart